+++ src/__tests__/conferenceParticipantTokenIssuedNotify.ts
@@ -1,3 +1,4 @@
+/// <reference types="jest" />
 import { createMediaStreamMock } from 'webrtc-mock';
 import type SipConnector from '../SipConnector';
 import { dataForConnectionWithoutAuthorization } from '../__fixtures__';