+++ src/tools/__tests__/callToServer.ts
@@ -1,5 +1,11 @@
+/// <reference types="jest" />
 import doMockSIPconnector from '../../doMock';
-import dataCall, { peerConnectionFromData } from '../__fixtures__/call';
+import dataCall, {
+  dataCallPurgatory,
+  onEnterConference,
+  onEnterPurgatory,
+  peerConnectionFromData,
+} from '../__fixtures__/call';
 import { dataForConnectionWithoutAuthorization } from '../__fixtures__/connectToServer';
 import parseObject from '../__tests-utils__/parseObject';
 import resolveCall from '../callToServer';
@@ -18,7 +24,7 @@ describe('callToServer', () => {
   });
 
   it('check call', async () => {
-    expect.assertions(1);
+    expect.assertions(3);
 
     return connectToServer(dataForConnectionWithoutAuthorization)
       .then(async () => {
@@ -29,6 +35,21 @@ describe('callToServer', () => {
         expect(parseObject(peerConnection._receivers)).toEqual(
           parseObject(peerConnectionFromData._receivers),
         );
+        expect(onEnterPurgatory).toHaveBeenCalledTimes(0);
+        expect(onEnterConference).toHaveBeenCalledTimes(1);
+      });
+  });
+
+  it('chould call correct handler after purgatory call', async () => {
+    expect.assertions(2);
+
+    return connectToServer(dataForConnectionWithoutAuthorization)
+      .then(async () => {
+        return call(dataCallPurgatory);
+      })
+      .then(() => {
+        expect(onEnterPurgatory).toHaveBeenCalledTimes(1);
+        expect(onEnterConference).toHaveBeenCalledTimes(0);
       });
   });
 });